update getPrice method to return float instead of int

+++ app/System/ShoppingCart/Item.php
@@ -22,7 +22,7 @@ class Item
         return $this->name;
     }
 
-    public function getPrice(): int
+    public function getPrice(): float
     {
         return $this->price;
     }
